pyproject.toml
[project]
name = "trading-strategy-generator"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.47.2,<6.0.0",
    "numpy>=2.3.3",
    "ollama>=0.6.0",
    "openai>=1.109.1",
    "pandas>=2.3.3",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
